% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QcPSPIntervals.R
\name{getListOfClimateVariables}
\alias{getListOfClimateVariables}
\title{Provide the Different Sets of Climate Variables.}
\usage{
getListOfClimateVariables()
}
\value{
a vector of characters
}
\description{
This function provides a vector that contains the
the different sets of climate variables as provided by BioSIM Web API.
}
