% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QcPSPIntervals.R
\name{getMetaData}
\alias{getMetaData}
\title{Provide the metadata for any of the data.frame objects provided by
the QcPSPIntervals and QcPSP packages.}
\usage{
getMetaData(dFrame)
}
\arguments{
\item{dFrame}{a data.frame object. Any of these objects: QcNbHarvestedTreesByIntervals,
QcNonoverlappingIntervals, QcSpeciesGrouping, QcTreeRemeasurements, QcClimateVariables,
QcPlotIndex, QcTreeIndex, QcMeasurementIndex, or QcTreeMeasurements}
}
\value{
a data.frame object containing the metadata.
}
\description{
Provide the metadata for any of the data.frame objects provided by
the QcPSPIntervals and QcPSP packages.
}
