% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QcPSPIntervals.R
\docType{data}
\name{nbHarvestedTreesByInterval}
\alias{nbHarvestedTreesByInterval}
\title{A data.frame object containing the number of harvested trees
by intervals.}
\format{
An object of class \code{data.frame} with 5074 rows and 4 columns.
}
\usage{
data(nbHarvestedTreesByInterval)
}
\description{
A data.frame object containing the number of harvested trees
by intervals.
}
\examples{
QcNbHarvestedTreesByInterval <- QcPSPIntervals::nbHarvestedTreesByInterval
}
\keyword{datasets}
