% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QcPSPIntervals.R
\docType{data}
\name{nonoverlappingIntervals}
\alias{nonoverlappingIntervals}
\title{A data.frame object containing the nonoverlapping intervals.}
\format{
An object of class \code{data.frame} with 37614 rows and 16 columns.
}
\usage{
data(nonoverlappingIntervals)
}
\description{
A data.frame object containing the nonoverlapping intervals.
}
\examples{
QcNonoverlappingIntervals <- QcPSPIntervals::nonoverlappingIntervals
}
\keyword{datasets}
