% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QcPSPIntervals.R
\docType{data}
\name{QcClimateVariables}
\alias{QcClimateVariables}
\title{A data.frame object containing the climate variables associated
with the nonoverlapping intervals.}
\format{
An object of class \code{list} of length 5.
}
\usage{
data(QcClimateVariables)
}
\description{
A data.frame object containing the climate variables associated
with the nonoverlapping intervals.
}
\examples{
data(QcClimateVariables)
}
\keyword{datasets}
