% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QcPSPIntervals.R
\docType{data}
\name{QcNonoverlappingIntervals}
\alias{QcNonoverlappingIntervals}
\title{A data.frame object containing the nonoverlapping intervals.}
\format{
An object of class \code{data.frame} with 37614 rows and 16 columns.
}
\usage{
data(QcNonoverlappingIntervals)
}
\description{
A data.frame object containing the nonoverlapping intervals.
}
\examples{
data(QcNonoverlappingIntervals)
}
\keyword{datasets}
