% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QcPSPIntervals.R
\docType{data}
\name{QcTreeRemeasurements}
\alias{QcTreeRemeasurements}
\title{A data.frame object containing the tree remeasurements.}
\format{
An object of class \code{data.frame} with 1479487 rows and 42 columns.
}
\usage{
data(QcTreeRemeasurements)
}
\description{
A data.frame object containing the tree remeasurements.
}
\examples{
data(QcTreeRemeasurements)
}
\keyword{datasets}
