% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QcPSPIntervals.R
\name{restoreClimateDataSet}
\alias{restoreClimateDataSet}
\title{Restore Climate data for Quebec PSP Data by Intervals in the Global Environment.}
\usage{
restoreClimateDataSet(climateVarSet)
}
\arguments{
\item{climateVarSet}{a set of climate variables}
}
\description{
This function restore a data.frame object that contains a
particular set of climatic data. The getListOfClimateVariables function
provides the different sets of climate variables as provided by BioSIM Web API.
}
