% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QcPSPIntervals.R
\docType{data}
\name{speciesGrouping}
\alias{speciesGrouping}
\title{A data.frame object containing the species grouping.}
\format{
An object of class \code{data.frame} with 64 rows and 5 columns.
}
\usage{
data(speciesGrouping)
}
\description{
A data.frame object containing the species grouping.
}
\examples{
QcSpecinesGrouping <- QcPSPIntervals::speciesGrouping
}
\keyword{datasets}
