% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QcPSPIntervals.R
\docType{data}
\name{treeRemeasurements}
\alias{treeRemeasurements}
\title{A data.frame object containing the tree remeasurements.}
\format{
An object of class \code{data.frame} with 1479487 rows and 42 columns.
}
\usage{
data(treeRemeasurements)
}
\description{
A data.frame object containing the tree remeasurements.
}
\examples{
QcTreeRemeasurements <- QcPSPIntervals::treeRemeasurements
}
\keyword{datasets}
